+++ src/components/organisms/DataTable/DataFilter.stories.js
@@ -2,7 +2,7 @@ import { storiesOf } from "@storybook/vue";
 import { object } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 
-import FilterMenuStandalone from "./FilterMenuStandalone";
+import DataFilter from "./DataFilter";
 
 import { tableData, tableFilters } from "./DataTable.data-factory.js";
 
@@ -12,9 +12,9 @@ storiesOf("6 Organisms/DataTable/SubComponents", module).add(
 		const data = tableData(10);
 
 		return {
-			components: { FilterMenuStandalone },
+			components: { DataFilter },
 			template: `
-			<FilterMenuStandalone
+			<DataFilter
 				:data="data"
 				:filters="tableFilters"
 				:activeFilters="activeFilters"
